Complete addition of animations
- restructuring of the sprites
- add enemy and player animation (grafics)
- adjust code
- add some more static vars in Controller.java

+++ src/de/halva6/snackman/model/Entity.java
@@ -26,12 +26,12 @@ public abstract class Entity
 
 	public abstract void move();
 
-	public double getPosX()
+	public int getPosX()
 	{
 		return p_x;
 	}
 
-	public double getPosY()
+	public int getPosY()
 	{
 		return p_y;
 	}
